+++ server/src/main/java/org/allaymc/server/block/component/button/BlockButtonBaseComponentImpl.java
@@ -139,9 +139,4 @@ public class BlockButtonBaseComponentImpl extends BlockBaseComponentImpl {
         }
         return 0;
     }
-
-    @Override
-    public boolean isPowerSource() {
-        return true;
-    }
 }