+++ server/src/main/java/org/allaymc/server/block/component/pressureplate/BlockPressurePlateBaseComponentImpl.java
@@ -175,9 +175,4 @@ public class BlockPressurePlateBaseComponentImpl extends BlockBaseComponentImpl 
         }
         return 0;
     }
-
-    @Override
-    public boolean isPowerSource() {
-        return true;
-    }
 }